feat(cli): extract swizzle into api/swizzle (receipt) + thin CLI
Move owner resolution, source copy, import rewriting, StyleX detection, and
maintainer-feedback building out of the swizzle command into api/swizzle.
`swizzle(name, opts)` performs the copy and returns a `swizzle.copy` receipt
(or `swizzle.list`); errors throw AstryxError with stable codes. The CLI only
parses flags, renders, and prefixes commands for the package manager.

Wiring: export `swizzle` from ./api + add to api.d.ts; move the rewriteImports
unit test to api/swizzle and add swizzle() list/not-found tests; promote swizzle
to a full api↔cli parity comparison (was a CLI-only probe).

Verified: swizzle --json output byte-identical (shape-lock: swizzle cases
unchanged; only ./api gains `swizzle`); typecheck:strict 0 errors; json-api;
packages/cli vitest 2073/2073; parity 330/330; json-smoke 15/15.
