+++ packages/cli/api/index.mjs
@@ -28,6 +28,7 @@ export {themeAdd, listThemes} from './theme/theme-add.mjs';
 export {hook} from './hook/hook.mjs';
 export {search} from './search/search.mjs';
 export {build} from './build/build.mjs';
+export {swizzle} from './swizzle/swizzle.mjs';
 export {doctor} from './doctor/doctor.mjs';
 export {layoutExpand, layoutCheck, layoutGrammar} from './layout/layout.mjs';
 export {
